fetch/metadata/generated/script-element-sec-fetch-mode-cors-script-element.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Mode - scripts via element</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(path, attributes) {
    const script = document.createElement('script');
    script.setAttribute('src', path);

    for (const [ name, value ] of Object.entries(attributes)) {
      script.setAttribute(name, value);
    }

    return new Promise((resolve, reject) => {
        script.onload = resolve;
        script.onerror = () => reject('Failed to load script');
        document.body.appendChild(script);
      })
      .then(() => script.remove());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        '' + path + '&mime=application/javascript',
        {"type": "module"}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - module script');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        '' + path + '&mime=application/javascript',
        {"crossorigin": ""}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - classic script, crossorigin=""');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        '' + path + '&mime=application/javascript',
        {"crossorigin": "anonymous"}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - classic script, crossorigin="anonymous"');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        '' + path + '&mime=application/javascript',
        {"crossorigin": "use-credentials"}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - classic script, crossorigin="use-credentials"');
  </script>
</html>
